Create magicNumbers.kt
03.2 Magic Dates
The date June 10, 1960, is special because when we write it in the following format, the month times the day equals the year:
6/10/60
Write a program that asks the user to enter a month (in numeric form), a day, and a two-digit year. The program should then determine whether the month times the day is equal to the year. If so, it should display a message saying the date is magic. Otherwise, it should display a message saying the date is not magic.
Write your code in a file named magicNumbers.kt in your unit03 folder. Turn in the GitHub link for a grade.